Move git module into `lib`
