+++ jobrunner/add_job.py
@@ -13,7 +13,7 @@ from urllib.parse import urlparse
 
 from jobrunner.create_or_update_jobs import create_or_update_jobs
 from jobrunner.lib.database import find_where
-from jobrunner.git import get_sha_from_remote_ref
+from jobrunner.lib.git import get_sha_from_remote_ref
 from jobrunner.log_utils import configure_logging
 from jobrunner.models import Job
 from jobrunner.sync import job_request_from_remote_format
