+++ jobrunner/github_validators.py
@@ -1,6 +1,6 @@
 from urllib.parse import urlparse
 
-from jobrunner.git import commit_reachable_from_ref
+from jobrunner.lib.git import commit_reachable_from_ref
 
 
 class GithubValidationError(Exception):
